Move redux store to store.js
